requirements.txt
pillow>=11.0.0
pillow-heif>=0.21.0
tqdm>=4.67.1
